Prepare medications stack to work with patient as “treatable”
Currently, if a medication is assigned directly to a patient, (i.e. not
assigned to an item such an “exit site infection”), the “treatable” role is null. There is not reason why the “treatable” cannot be assigned to the patient, there it would mean a medication requires a treatable association.
This means direct patient medications can work with the polymorphic
medications controller setup for infections and episodes.