Fix failing reviewing_clinical_summary feature
This cucumber feature is failing due to a change in
the Prescription.default_sort_order from desc to asc.
Add a change to ensure prescriptions in cucumber hash tables
are created with an appropriately incrementing `prescribed_on`
date so that when the results are verified, they will be in the
order specified in the Given prescriptions hash table.

Also modifiy circle.yml to specify a progress formatter
so we see cucumber output for non-web tests. We also have
to add in the rspec tests, becuase now we are overriding
circleci's deduced tests, we need to be explicit about what
we want it to test.