Add BM Stix validator and refactor weight and temperature validator
There is a lot of commonality between the validators i.e. a range and scale (dp).
Break out concerns to share between the validators.
Keep specs for each validator but try remove variations so refactoring is easier later.
No need to refactor to shared examples yet as this makes them harder to read.
Once we get to four similar validators, shared examples might be the best approach.