HD Session validation
A first attempt at validation by status.
If attempting to close the Session, we cast it to a Session:Closed and validate it. The Session::Close class is a bit of a mess- the embedded SessionDocument does not lend itself easily to conditional validation.
Having tried several approaches, this one is the least complex but does not adddress
- the conditional validaion of HDF fields
- the presence validation of BloodPressure value types
I think a better solution (offering more control and clairty over validation  now and in the future) is to use Form Objects but this may involve more duplication and wil be a bit more long-winded.