Rename sass partials to start with _
This is a sass convention to communicate to a developer and
to the sass precompiler that the file is a partial
for import elsewhere and not to be rendered to a discrete css file.
Also remove the .css. element any .css.scss extensions as they
aren’t neccessary and doesn’t communicate anything.