Remove spurious I18n data in HD Session view
Because of the way I18n AlwaysCascade works, it will look up the tree for a missing key but will not find a top level one that only exists in en.yml - instead at the top level it only resolves keys in en-GB.yml (if that is out locale). In the case of this bug, the :true and :false keys were only defined in en.yml and so were not resolved, causing a hash the complete hash of en-GB 1i8n keys to be spat out instead. I.e. the solution is to move any keys that cause this problem from en.yml into en-GB.yml into en.yml