Coerce glucose_content to a string before querying db
Coercing glucose_content to_s because where(glucose_content: 1.36) returns no rows
even though the raw query in the log, run in pgadmin, works fine. AR doing something odd.
Since migrating to Rails 5, querying the glucose_content column *for some reason*
won't work using a float on the RHS. Works for sodium_content, not for glucose_content
even though they are similarly defined. IN fact sodium_content is numeric(3,2) and
glucose_content numeric(4,2) but changing it numeric(3,2) doesn't help. Its odd.
Hence the #to_s which means I presume that AR will not get to misinterpret it and
it will leave pg to query with it and get it right.