RTF rendering - use pandoc
Use pandoc to convert html to rtf if the a formatted_letters url with the .rtf extension is requested.
Pandoc must have been installed using hombrew or apt-get etc.
Note that although RubyPandoc (a thin pandoc wrapper) has an API which allows you to pass an html string (for example from `render_to_string`) this produces a different output to using pandoc with a physical html file. So we save the html content to a tempfile, convert that to rtf (rubypandoc captures the output) and return that using senddata. Note rtfs cannot be displayed in the browser so there is no concep of disposition: inline, only disposition: attachment.