Fix Rails 5 letter issues
Set Draft letter STI type manually. This is a hack to addresses odd errors
where the Rails STI mechanism is not setting the type automatically when
you create a new Letter::Draft. We need at some point to get to the bottom
of this. Note that you need to run all letters features in order to start
seeing the errors e.g.
  bundle exec cucumber --format progress features/renalware/letters
Running individual failures works fine - this suggests its a dependency
load order issue perhaps?