Generate and store a base58 token in secure_id and use in patient urls
Use the built-in rails has_secure_token method to populate a new secure_id column with a 24char
base58 token, and we use this (with the help of the friendly_id gem) in patient urls.
In pg the patients.secure_id column is defined as not null with a default value that calls a
custom pg function to generate a base58 token (almost identical to the rails implementation of
this function) as patients could be created outside of rails in which case the secure_id
would otherwise not be populated.