Allow clicking on the clock icon to display the clock picker
There was a bug (I believe) in the open source project which was stopping a click on the optional .input-group-addon (which contains the clock icon) from opening the clockpicker. I have moved the dependency on the original clockpicker gem and moved the 2 files into our app/assets folder so we have explicit control over them and I can fix the bug without worrying about the change being overwritten.