Add Liquid template rendering
Allow a template to be loaded from the database, merged with a sepcific subset of data exposed using Liquid Drops, and rendered for example from a controller action.
This lets us seed hospital-specific one-off html views (for example the Barts esi form) that a hospital can design but which can still pull in a very limited amount of  developer-defined patient data.