Delegate patient#diabetic? to patient.document.diabetes
We currently store diabetic in 2 places - on Patient directly and in document.diabetes where we have a boolean diagnosis (which might become an enum of Type 1 or Type 2 in the future) and a date of diagnosis.
This change removes the Patient.diabetic column which is not used very much atm, and instead Patient.diabetic? now delegates to document.diabetes.diagnosis