Create mechanism for Audit management and viewing
The mechanism is based on there being a materialized view behind the audit
(with a name starting with `reporting_`) , which
is refreshed via cron at a schedule specified by audit#refresh_schedule. Data is displayed
using jQuery DataTabes, configured by JSON in audit#display_configuration and data retrieved via
an ajax call to e.g. `reporting/audits/1.json` which selects * from the materialized view and
and converts to JSON.
Drilling down is not yet implemented, and will follow as another commit.

- Add Audits model and controller and reporting/audits route
- Add DataTables gem
- Display audit data in DataTables by querying the associated materialised view and converting to JSON
- Allow configuration of an Audit via a form where a use can cnahe the name,, description, databales config etc
- Add abilty to refresh the materialized view behind an audit using an ActiveJob
- Add a pg fn to refresh all materialised views and call this from the dummy app’s seed file once all data seeded - otherwise the MVs will be empty
- Add (WIP) Main Authors, HD Blood Pressure and HD Overall audits as demonstrations - all need some work and are not intended to be the finished thing