Use flash.now[:error] in actions when rendering rather than redirecting
If we do e.g.  after a validation error in an #update action, we should use  otherwise the flash message stays around for the following action - e.g. if they have an error on PageX and then navigate to PageY, currently they see  on PageY (an unrelated and confusing message). Using flash.now means the flash message is only around for the current action (they one being rendered).