On Death modalty, terminate drugs and don't cc letters to patient
- Create new ChangePatientModality object to start to decouple Rewnalware::Patient from the
Modalities module, and to provide a better place from which to broadcast modality
change events. Easier to test also, and removes god-like behaviour from Patient.
- Add new means of subscribing to a globally-defined (in Renalware.config) list of interested
subscribers (listeners). The existing SubscriptionRegistry was not working across threads in Puma.
- Broadcast an event using Wisper when a patient modality has changed
- Listen for Death modality change events in the Letters module and in answer set the flag to
indicate the patient should no longer be CC'ed on letters
- Listen for Death modality change events in the Medications module and in answer terminate
all active prescriptions for the patient
- Add js alert to warn that Death modality clears current prescriptions