Normalise HD schedule
Change HD Schedule from a string to a model called ScheduleDefinition which can be assigned
to a patient's HD Profile or PreferenceSet, and allows us to use a SQL query to more flexibly
find patients dialysing on a particular day and period (e.g. am).
See comments in the ScheduleDefinition class definition.