Switch from base58 to uuid for patients.secure_id
This is to get around duplicate secure_id issues during data migration. The uuid format is a bit longer and less ‘secure’ but sufficient for our needs. The upside is that it is slightly faster to search by uuid in the database, and they require less storage.
Note that as far as postgres is concerned, when for instance searching by a patient’s secure_id,  a12d9a8e-9cc9-4fbe-88dd-2d1c983ea04f is the same as a12d9a8e9cc94fbe88dd2d1c983ea04f i.e. hypnens are optional.