Rename Inpatient -> Admission
The name of the Inpatient model seemed to not quite work in the app i.e. a patient having many inpatients or an inpatient belonging to a patient. Instead I have reveted to a previous suggestion in #1568 to an Adimission model so a patient has many admissions.
The downside is we have Admissions::Admission, but we also have e.g. Letters::Letter so at least its consistently awkward!