Change ObservationSetMethods.method_missing to allow any *_result _observed_at message
Don’t restrict to allowing only requests for codes defined in Pathology::ObservationDescriptions.
This is because the new grey clinical bar loads on every feature spec and checks for eg hgb_result and if HGB is not pre-defined in ObservationDescriptions it would blow.
I think this is a more flexible solution and low impact.