Fix bug where devise sign in timestamps not being saved
Because users with no profile or signature are ‘invalid’ when trying to save, devise was not able to update the last_sign_in_at and current_sign_in_at timepstamps on the user row.
Have changed the logic to make position and signature non-mandatory unless you are in the the Edt Profile form.
We need to move the profile and signature to a UserProfile model at some point to clean this up.