Keep the user’s search string intact
We remove spaces from nhs number etc before searching. Previously we replaced the users search string with the sanitized version (e.g. the version with spaces removed from the nhs number) but I think this is confusing for the user and unnecessary so now we keep their original search string intact when re-dsplaying it