Remove unique obr requestor_order_number
Turns out when there are > obr in a message, each OBR has the same requestor_order_number. So we need to allow duplicates. When exporting to UKRDC we need to find out whether laborder == obr or laborder == ‘requestor_order’ which could have many obrs. If the former, need to challenge this as seems incorrect, or add a the OBR count to the end of requestor_order_number to keep UKRDC happy