Display practice address when searching for a GP
Now displays practice address so you can see what matches
If searching for Maddock, will find
- practice called paddock
- first line of address with maddock in it catches the case where recep could not find Maddock Way Practice where name of practice was actually Dr Isak’s Practice and MAddock way is the first line of the address!

Searcing by postcde works

Not you can search by GP name first and then search for the practice. Not sure if that is going to be a problem.

Limit results to 50 so searching for Surgery does not bring back a thousand results