Allow soft delete of consults with no type column
Migrated consults may have no type, so when they are marked as Done (ie soft deleted) they updated rather then destroyed (ie ended_on is set) and this causes validation to fail as there is no `consult_type` value assigned to the consult. User are not sure what the type is for but we’ll keep it for now and only validate on create, not on any updates.