Cast letter to Letters::Letter class before emailing
If we add a letter of Letter::Approved class to the delayed_job queue, and before it is processed someone prints it (changing its state to completed) then when the job kicks in tries to load the letter using say Letters::Approved.find(100) it won’t find it the class has changed, and it won’t get emailed.
For this reason we must cast the letter to its superclass `Letters::Letter` using `becomes` before adding it to the queue.