Allow email meta data insertion when patient has no practice
We use the same PracticeMailer (possible misnomer) to send emails to Practices and to QEH admin. This always inserts meta data into the email body, and this contains GP and Practice info. We currently allow a missing GP in this meta data but not a missing practice - its a PracticeMailer afterall. But as mentioned we use it for QEH patients and those might not have a practice or GP). So we need to allow those to be missing when creating the meta data.