Make HIV HepB and HepC accept a status of Y N U and a year
Note cannot chnage diabetes to Yes No Unknown + Year because current patient document jsonb stores true/false and a date for Diabetes. Will need to write a migration to update in the jsonb,  or some code to handle the display/form when the diabetes is not in the expected format - tricky, and one of the (relatively few) downsides of the jsonb approach.