Add a unique MD5 hash of HL7 message to feed_messages
If the incoming HL7 message has already been received and processed we should not be processing it again. To help enforce this there is a unique MD5 hash on feed_messages which will baulk if the same message payload is saved twice - in this case we exit call early, the broadcast
is not issued and therefore the message is not processed. The message will go back into
the delayed_job queue and retry, failing until it finally gives up