Fix bug where invalid jsonb dates prevent form loading
If we see a string when loading the datepicker input we can assume it is an invalid date eg 12-00-2010 that has for
example been imported from another system, and its one we store in Rw2 in jsonb where there is no database type checking,
In this case don’t try and format it to the locale format.
Have seen this in LowClerance::Profile.document.