Fix f.object issues
Pass letter variable into partial rather than use f.object which in this instance (as of Rails 5.2) was not taking with it the letter decorated by the presenter, but just the underlying letter. Unclear why, but moving away from the use of f.object is an improvement in any case.