If present, display external study link in study alert on patient pages
When displaying an alert on patient pages to indicate they are a participant in a study, if the study has an external application_url, display a link in the alert. When the user clicks this, open up the application in a new tab, landing on the partcipants page within the application.