Add StudyParticipantsController#show action
Certain cirumstances are causing a StudyParticipantsController#show to be invoked but this path does not exist and we don’t support it - so now if it is called we redirect to StudyParticipantsController#index to display the list of participants for the study.
Not sure yet if the offending call is format `js` or `html`. If we see this error again we will know.