Removed js copy to clipboard for now as UI not right
For now in Chrome and Firefox only it selects the entire hosp/nhs number (using css) when clicked on making it easy to then copy the number.
Later we will add some js to add the similar functionality to IE/Egde.
I have removed the automatic copy to clipboard as felt the UI was not quite right it it deselects anything manually selected when clicked on.
Not a big job to get this right but leaving for now - this PR as it stands will provides some help for Chrome users.