Refactor study participants to remove use of a modal
A modal was previously used to add/edit a participant. We need more screen real estate now that we are extending what can be added to a participation, so move back to using a regular page.
This commit also brings in the capybara-select-2 gem which is a better implentation for testing select2 than we had previously.