Rename StudyParticipant -> Participation
This better suits the nature of this cross reference table and makes the work of enhancing studies (in another PR) somewhat easier.
Also rename participant_id -> patient_id on study_participations to make the relationship clearer.

Note the trigger on study_participations that generates a pseudo-encrypted external_id will
continue working as Rails kindly re-wires it during the migration. The name of the trigger and the
function it calls still use the name 'participant' but we can live with this for now.

- Add user memberships to research studies
- Restrict Research edit/create/destroy access to admins and superadmins
- Make tabs in Research studies non-js driven - Drive them manually for more control
- Rename study members ‘investigators’ - Also use the awful name ‘investigatorship’ to replace ‘membership’
- Resolve some git rebase issues
- Add a ClassFactory to deduce the Study/Participation/Investigator concrete class
- Store participation STI type - Participant class is determined by the namespace setting on Research::Study. Ensure we capture the STI class of new participations
- Store investigatorship STI type
- Add jsonb douments to Study, Participation, Investigatorship
- Support custom forms for Study, Participation, Investigatorship - A host app can define a study and its custom form partials that can be used to save/edit custom data.