Broadcast event_created event_updated messages
When an Event is created, broadcast an event_created message, passing the event, so that a listener (eg the host app) can decide what to do with the event - for instance it might create and store a PDF version of the event.
When an Event is updated we broadcast an event_updated message.