Create a PDF from an Event
We already broadcast and event when an Events::Event is created or updated so that a host
application can intercept this. Here we support generating a PDF version of the Event so that the
host application can choose save the event to the electronic pubic register (EPR) after creation.
To aid this use case we provide a new job - SavePdfEventToFile - which in the background renders the
event to a PDF and saves it to the requested location, e.g. an EPR folder.

We use prawn to render the Event PDF since this is no existing HTML representation to render with
wicked_pdf/wkhtmltopdf. It is also considerably faster.

To support adding hospital centre name, address and contact info to the Event PDF, we add new
columns to hospital_centres. While letters use Letters::Letterheads for storing the address, contact
numbers etc, Events are in an unconnected domain and cannot re-use Letterheads. So we have to
duplicate (and at some point migrate) site info from letter_letterheads into hospital_centres. This
is also in fact probably the best place for this info.

The Hospitals::Centre whose info we put in any Event PDF is the one found whose code matches
`Renalware.config.ukrdc_site_code`. An error is raised if this is not set or not found.

Improve event type seeds to add save_pdf_to_electronic_public_register.
Improve hospital seeds to add info, trust_name, trust_caption.