Support for flexible and extensible grouping of pathology results
Previously we had 2 fixed groups - ‘display’ and ‘letters’. This PR  allows us to have multiple specific sets of OBX codes for use in certain contexts eg letter, HD MDM etc.
Members within a names group (eg ‘letters’) have a subgroup and position within the subgroup.
This initial PR merely creates the tables and relationships, so that once deployed we can populate the data ready for a future PR which wikll switch over to using them in place of the existing group mechanism.