Add Respiratory Rate to HD Session
Add pre- and post- respiratory rate measurements to the HD sesssion form. Validate respiratory_rate is an integer in the range 4 to 40.
Include a respiratory_rate_measured radio button that, if No, skips validating the presence of respiratory_rate.
This allows automated session ingestion to set this to No when the data is not available in a feed.