Migrate from CircleCI to SemphoreCI
CircleCI is slower because you have to pay for parallelism, and also we
found that if you use 100% of your free minutes, you are left having to
pay $50 for the last few days of the month.

SemphoreCI has a business model which suits us better: unlimited parallemism
from the start of your free minutes, then pay as you go, which brings
faster tests (4.5min as opposed to 11min) and is cheaper in the long
run.