Add authorisation to HD Drug administrations on the HD Session form
Enhance the drug administration section in the HD form so that, for
each drug:
- If the drug is administered, a witness and nurse must 'sign' by
entering their respective user and password
- Default the Nurse user dropdown to be the currently signed-in user
- When entering a password, authenticate the entered user/password
immediately on blur and return a token which will be submitted in the
form when saved. This token indicates the authentcation was successful
and need not be done again. Details on how the token is generated TBC.
- Up to the point the session is signed-off, it is possible to clear the
authentication so the drug must be re-authorised.
- When marking a drug as administered it highlights it in red until it
is fully authorised. Once signed it highlights it as green.