Add SQL view to resolve the hd_profile associated with an HD modality
If there is no profile active by the end of the day on which the modality is created, it looks aread 14 days to find the first one created within that period. If there are several profiles created on one day it choose that last newest one.
The view allows us to eyeball what profiles are being used for UKRDC export; HD modalites with not associated HD Profile will not generate a complete Treatment because we cannot tell what the hospital unit or hd type is (hdf etc).