Support legacy html views in host app via a ‘bottom’ partial
We used to override a Renalware engine view in the host and call renalware_view to render the (super) engine view from there. This led to poblems with javascript not executing.
This alternative approach so to attempt to render a _bottom.html.slim partial in places where pulling in legacy views in the host might be needed. If no _bottom partial found nothing is rendered, otherwise it will render the partial.
This gets around javascript load order problems.