Allow a sire to disable HD session drug signoff
Add a configuration option to disable requiring HD prescription adminstrations to be authorised (by a nurse and witness) when completing and signing off an HD session in the HD Session form.
Hide the nurse and witness username/password fields but
keep ‘Reason not given’ dropdown which can be completed if the drug was not administered.