Do not keep failed UpdateRollingPatientStatisticsJobs
Because this job will be triggered again the next time an HD Sessions is created, it is not crucial to keep each event around - ie they have a short shelf-life. For this reason we only retry 3 times then delete failed jobs.
Note after migration at a site, the last 12 HD Sessions for a patient might trigger various errors as they have not been created through Renalware, so it is expected for some of these jobs to fail in the first month or so.