Ignore duplicate HL7 messages
Occasionally a duplicate mesage arrives. We know it is a duplicate becuase we calculate and store a hash of the body of the message in the feed_messages table; if another mesage arrives with the same hash we can safely assume it is a duplcate. Rather than let duplicate index violations fill up the delayed_jobs table, handle the error and log it as a warning.