Ensure problems and notes are ordered correctly
- add an order to patient.problems and problem.notes relationships
- refine Problem.with_notes scope to eager load notes and order them correctly by created_at asc

There is still an issue with the position counter in problem_problems (used for sorting problems) incrementing across all problems, not just within a patient’s problems, but mysteriously I cannot find where the SQL max(position) call happens - not in the app or in the database. It does not appear to be a trigger, or a callback or whatever.