Fix model sorting by #position issues
Fix a global issue with sorting - currently when adding a new row to a table that has a ‘position’ column, the new position is the max(position) + 1 across the whole table. Change this so that if a .position_sorting_scope scope is present on the model, the new position can be limited by the scope to for example be max(position) where patient_id = 123. This means each row will in this case will have positions ascending cleanly as 1,2,3.. per patient. It will also prevent massive position numbers building up for example in the problem_problems table.