Begin conversion from Cucumber to Turnip
Having cucumber as separate process to run when testing and on CI, and as a
separate 'technology', causes increased
complexity and developer cognitive friction. In addition cucumber has arguable become
less popular and thus future developers will have less motivation and
abilty to support it. Being able to use RSpec throughout
the test suite is a distinct advantage in running and managing test
code. In this PR we start to migrate from cucumber to turip which allows
us to continue to use gherkin syntax (Given Wehn Then etc) but include
thenm in the RSpec test suite. There remains a small fly in the ointment
- because the same features can be flagged as non-javascript and as @javascript ('web')
we need to spawn a separate RSpec runner to execute the second pass. Once updated
to Rails 6 we might be able to do this in parallel.

As part of this refactor we consolidate the use page objects and
factory_bot factories so these are shared across all tests, and move
away from the fixture-based approach we currently use in the cucumber
features.