Handle ActiveRecord::PreparedStatementCacheExpired errors during deploy
Occasionally a user might experience an ActiveRecord::PreparedStatementCacheExpired error while a new version is being deployed is the there is   a migration which changes one or more columns used in a cached PostgreSQL prepared statement.
If we get this error while creating a letter or HD session. retry the transation once before accepting defeat and re-raising the exception.