Allow skipping validation of died_on and first_cause
If the patient modality is Death we normally validate on update that they have a cause of death and a died_on date.
Add a boolean virtual attribute `skip_death_validations` that when set to true will cause validation of died_on and first_cause_id to be skipped. This allows non-interactive process such as ADT import to change things about a deceased patient and still be able to save those changes even the patient is missing a died_on and first_cause_id.