Remove webpacker
Have switched to Sprockets 4 as a way to start using es6 and stimulus in Renalware. The addition of webpacker in the engine is a pain and I think in any case we need to be using something like rollup.js to build an npm (or github package) js bundle to include in the host app rather that the hackery that webpacker-in-an-engine requires e.g. injecting the engines pack into the host at compile time.