Fix a bug where author is not being remembered between letters
The first time a letter is created after logging in, the author dropdown should default to the current logged in user. Successive new letters (until the user is logged out) should automatically select the last chosen author, which may or not be the current user.
Create a system test for this and other remembered letter choices.