Do not allow a future modality start date
When adding a new modality to a patient, do not allow a date after today (ie in the future). This is to prevent an accidental future date being added - a mistake it is not possible to correct because subsequently added modalities can only have a start date after the previous one.