Support mutiple homecare supply providers
When printing ESA and Immunosspressant prescriptions (and
potentially other drug types), allow a different Provider and
their printable PDF form to be assigned to a given drug type.

For example ESA homecare drugs could populate and launch the
Fresenius PDF form version 1.

PDF homecare forms are defined in external gems.
- renalware-forms provides a generic homecare supply form and the
  API to populate any home care form
- renalware-forms-fresenius implements the Fresenius PDF form and
  can be populated using the API provided in renalware-forms
- renalware-forms-* gems for other providers

We use stimulusjs controllers to manage the modal dialog used for
selecting drug type and other variables that are fed to the
homecare form