Fix pathology sorting issue on for tests on the same day
Fix an issue in Historical and Recent pathology - when the same test for a patient ocurrs >1 time on the same day, the most recent result result was not always shown because the order was undefined. Make it so the most recent test result is guaranteed to be displayed.
Note that in this fix that, while the addition or the `order by observed_at asc` in the where clause seems unintuituve, it does return the correct result.