Add hidden boolean attribute to User
If a user is flagged as hidden in the admin/users UI then it should not be possible to assign to them as
- letter authors
- electnoric cc recipients
- messages
Need to be careful not to apply this logic to places where models can be edited - for exampe if you assign a user to a consult and then make the user hidden, when you edit the consult you would presumably want to see the previously chosen user pre-selected in the list (even if they are hidden). Some thought required how to handle this instance.