Support configurable pathology code groups
Allow a superadmin to cofigure which pathology observation descriptions (OBX codes) are displayed in which subgroup, and in what order. These groups are named (the name is fixed and not user-editable) and used anywhere a predefined set of codes need to be displayed. For example on the HD Protocol we display the 'hd_session_form_recent' group and the hospital can define which set of codes are displayed wherever this group os displayed.