Fix confusing kaminari paging issue in HD MDM list
The count used to assess pagination totals/pages was using 2 modality_modalities joins rather than just the one that the actual data uses causing a count mismatch with the number of rows displayed.
We may move to using SQL views for complex aggregate tables to avoid this sort of complexity.