Refactor HD prescription administrations
These now belong to a patient and not a session.
Using a dialog the user can choose the drug and whether
it was was administered or not. If administered, they can
enter their name and password. If the witness is present they
can enter their credentials also. If witnessing later they can still
save their credentials, and a witness can later on enter their password
(or change the witness name if they were assigned incorrectly) in a
simplified dialog.

Also
- Ensure autoprefixer supports IE 11
- Make Admissions and Consults headings into hyperlinks to be consistent with other hyperkllinked headings