Fix bug clearing UKRDC data when patient dies
When a patient dies the patient model enters a state where no
further changes can be made to it unless you set the date_of_death
and death causes. However this validation can be overriden by setting
the virtual accessor Patient#skip_death_validations = true.

This commit adds this to ClearPatientUKRDCData to prevent it
raising a validation error when it attempts to save
the patient.