Catch PDF encoding errors before an Event is sent to EPR
Events are rendered to PDF using Prawn, using which creates lightweight
files (5 to 6KB) because it relies on the PDF fonts available in
PDF viewers and does not embed any custom fonts, which would take the
file size to > 40KB. The built in fonts appear only to support the
smallish Windows-1252 character set.
If a  user pastes an unusual character into the Event description or
notes and saves, and the character cannot be converted into
UTF8 (most unusual chars can) we now display a validation message asking
them to remove those characters. Most unusual characters will be accepted
and displayed as gibberish, which is fine, but some cannot be converted
even into gibberish and would otherwise raise an error when Prawn
asynchronously generates the PDF.