Introduce tailwindcss
Long term we'd like to remove Foundation 5 in favour of tailwindcss.
Upgrading to Foundation 6 is another option but is in itslef significant
and I am not sure about Foundation as a framework anyway. Tailwind encourages
more UI tweaking and customisation and I think will let us take the UI to a
new and modern level. Introducing Tailwind does however involve some new
moving parts. It is an npm package and requires webpacker or in our case
rollup.js to compile the css and put it into a .css file which the
asset pipeline can include. In a staging or production environment we
also need to make sure the compiled css is run through purgecss (via
postcss) in order to reduce the filesize frim about 1.6mb to a hundred
kb or so. This is done using a postinstall script task in package.json.
Running yarn install for instance will install the packages and then run
the script which will create rollup_compiled.css ready for inclusion.

Changes of note
- Rename existing css class grid to mgrid To avoid a clash
  with the tailwindcsss grid utility class
- Add @tailwindcss/custom-forms npm package which provides better
  default styling of forms
- Ignore rollup_compiled.css - it is generated on install by rollup from
  a postinstall script in package.json
- Note I could not find a way to invoke purgecss etc
  whle keeping it (and rollup, postcss etc) in devDependencies
  in package.json - so there are a lot of new
  non-dev dependencies. Will look into a way around this.