Keep a user's session alive while during activity in the browser
For example when the user is typing a letter there will be a stream of
keypresses which should periodically signal to the server that the session
is alive. This prevents the session expiring and the page redirecting
to the login page during a letter that is taking a long time to write.
The same applies to clicks, but avoid scroll or resize events
for now until we understand any performance impacts a bit better.
Re-implement the current session expiry polling logic but adjusted to
avoid unneccessary polling and use a stimulus controller.
Log out other tabs immediately when user logs out manually or is
timed-out, communicating with other tabs using a localstorage event.