Simplify UI and SQL views by removing tabs
Remove tabbed 'filters' like 'On worryboard', 'Tx within 3m' etc
and make these boolean columns returned by the view,
and add filters so a user can choose Yes or No in the
On Worryboard dropdown filter for example.
Previously the tabs where implemented as descrete SQL views
which query a parent view ( ie 'All'), adding a where clause
and perhaps joining onto patient _worries etc. The downside
of encouraging this stacked view-upon-view approach is that
if you alter/add a colum to the SQL view behind the  'All' tab, you
have to drop and recreate all views that depend on it, making it
messy to work with.
So it is less work to have one large view that builds in all
possible filter/selection options.