Capture PD PET/Adequacy completion state
Derive the #complete attribute based on the presence of certain calculated fields.
An Adequacy may be complete even if some calculated fields are null, provided an appropriate
urine-missing or dialysate-missing boolean is set. These indicate that the patient's urine
or dialysate samples where missing and consequently renal calculations (relating to urine) or the
peritioneal calculations (relating to dialysate) could not happen.