Fix missing patient rule when printing one form
If you create a patient pathology request rule with freq Once
and print the form, the rule is missing on the PDF.
This seems to be because in the process of printing,
the rule is 'invoked' prematurely
thus disappearing in the final print .
The fix is to use momoisation to cache the
results rather than reloading twice
during printing. This is a general
improvement in performance anyway.