Add a numeric  nresult to pathology_observations
Add an nresult column of type float to pathology_observations.
Add a BEFORE UPDATE INSERT trigger on pathology_observations
that will call a function that will
attempt to set nresult to result::float, and if the cast
cannot be made, use a NULL for nresult.
This pre-casting of result to a float enables performance
optimsations for example when fetching numeric graph data;
non-numeric data points can easily be excluded without a runtime
cast to a numeric value.

Decision to use float rather than numeric(x,y) is for performance
as numeric is stored as BCD and somewhat slower if we use nresult
for statistical analysis, given there might be 100m rows.

Float8 us used rather than float4, which, though having the
advantage of taking half the space, displayed some
unusual rounding affects I did not understand.