Only display public patient messages
Introduce a #public attriobute on internal messages. This defaults to false for historical messages.

New messages are created with `public: true` and the user should see some info on the
New Message form to indicate their message will be viewable by all.

Only display public messages on the clinical summary and patient's messages page.

This is because up until now a user might written a message which was intended to be private.

Private messages are not longer possible.