Heroku app.json changes to fix broken builds
- Sendgrid have give us the cold shoulder - possibly due to us using
  them for review apps for years on the free plan? - and so we have
  switched to the free mailgun addon for email (password resets etc)
- Remove the deprecated engines section
- Specify postgresql version 11
- Crucially change bin/rake to require "../spec/dummy/config/boot";
without this, a recent change - possibly at Heroku as I cannot find
anything locally - Heroku PR builds fail with 'cannot detect rake
tasks' while failing to load config/boot fron the wrong location.
It seems it has started to target the root bin/rake rather
that the one in spec/dummy/bin that was being used before, and from the
root of the app there is no ./config/boot.rb file as it is an engine.
I am still unsure why this has stated happening or whether my change will have any
knock-on effects. I tried an older versions of the heroku/ruby buildpack
and had the same issue which is odd, though it is possible some caching
at Heroku means I was still using the new one.
Its possible its to do with something in Heroku
that sits around the ruby buildpack. Or it could just be something silly I
have done. I am not sure about this though as when I creating a review app from a 6
month old pull request I still get the error. However I suppose it is
possible that there is some global config in the Heroku pipleline config that has got
lost - something that was steering Heroku to use ./spec/dummy/bin/rake and
not ./bin/rake.